moved all functions to make npm work
